requirements.txt
streamlit
ccxt
pandas
pandas_ta
plotly
flask
openai
python-dotenv
requests
pyTelegramBotAPI
